add stream manager
